+++ src/main/java/com/cloudDomus/cloudDommus/Service/ServiceController.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import org.springframework.transaction.annotation.Transactional;
 
 @RestController
 @RequestMapping("/api/services")
@@ -33,13 +34,25 @@ public class ServiceController {
     @GetMapping("/service/{id}")
     public Service getServiceByID(@PathVariable Long id) {
 
-        return repository.findById(id).orElseThrow(() -> new ServiceNotFoundException(id));
+        return repository.findByID(id);
     }
 
     @ApiOperation(value = "Delete a service by its Id", response = List.class)
     @DeleteMapping("/Service/{id}")
-    public void deleteService(@PathVariable Long id) {
+    @Transactional(rollbackFor = ServiceNotFoundException.class)
+    public Service deleteService(@PathVariable Long id) throws ServiceNotFoundException{
+        Service deleted = repository.findByID(id);
+        if(deleted == null){
+            throw new ServiceNotFoundException(id);
+        }
         repository.deleteById(id);
+        deleted.setId(null);
+        return deleted;
     }
 
+    public void setRepository(ServiceRepository repository) {
+        this.repository = repository;
+    }
+    
+    
 }
